--- src/main/java/com/alertops/dto/escalation/EscalationStepDto.java
@@ -1,32 +1,0 @@
-package com.alertops.dto.escalation;
-
-
-public class EscalationStepDto {
-    private  Long userId;
-    private  String userName;
-    private String email;
-
-    public Long getUserId() {
-        return userId;
-    }
-
-    public void setUserId(Long userId) {
-        this.userId = userId;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
