--- src/main/java/com/alertops/flow_execution_engine/controller/ScheduleController.java
@@ -1,29 +1,0 @@
-package com.alertops.flow_execution_engine.controller;
-/*
-package com.alertops.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import com.alertops.dto.schedular.SheduleReq;
-import com.alertops.service.SchedularService;
-
-import java.util.List;
-
-
-@RestController
-@RequestMapping("/api/v1/scheduler")
-public class ScheduleController {
-    private final SchedularService schedularService;
-
-    public ScheduleController(SchedularService schedularService) {
-        this.schedularService = schedularService;
-    }
-
-    @PostMapping("/start")
-    public ResponseEntity<?> scheduleBulk(@RequestBody List<SheduleReq> list) {
-        schedularService.startScheduleBulk(list);
-        return ResponseEntity.ok().build();
-    }
-}
-*/
