+++ src/main/java/com/alertops/task/task/UserTasksResponseDto.java
@@ -1,4 +1,4 @@
-package com.alertops.dto.task;
+package com.alertops.task.task;
 
 public class UserTasksResponseDto {
     private Long id;
